+++ src/KardFinancial.Test/Unit/MockServer/Organizations/ContentStrategies/CreateTest.cs
@@ -20,6 +20,11 @@ public class CreateTest : BaseMockServerTest
                 "attributes": {
                   "name": "Featured Travel",
                   "sort": "HIGHEST_CASHBACK",
+                  "filters": {
+                    "offerFeatures": [
+                      "INTERACTIVE"
+                    ]
+                  },
                   "categories": [
                     "Travel"
                   ],
@@ -42,6 +47,11 @@ public class CreateTest : BaseMockServerTest
                 "name": "Featured Travel",
                 "organizationId": "org-123",
                 "sort": "HIGHEST_CASHBACK",
+                "filters": {
+                  "offerFeatures": [
+                    "INTERACTIVE"
+                  ]
+                },
                 "categories": [
                   "Travel"
                 ],
@@ -81,6 +91,10 @@ public class CreateTest : BaseMockServerTest
                     {
                         Name = "Featured Travel",
                         Sort = ContentStrategySort.HighestCashback,
+                        Filters = new ContentStrategyFilters
+                        {
+                            OfferFeatures = new List<OfferFeatures>() { OfferFeatures.Interactive },
+                        },
                         Categories = new List<CategoryOption>() { CategoryOption.Travel },
                         CategoryExclusions = new List<CategoryOption>() { CategoryOption.Gas },
                         MerchantExclusions = new List<string>() { "merchant-abc" },