+++ src/KardFinancial.Test/Unit/MockServer/Organizations/ContentStrategies/GetTest.cs
@@ -19,6 +19,24 @@ public class GetTest : BaseMockServerTest
                 "name": "name",
                 "organizationId": "organizationId",
                 "sort": "NEWLY_LIVE",
+                "filters": {
+                  "categories": [
+                    "categories",
+                    "categories"
+                  ],
+                  "categoryExclusions": [
+                    "categoryExclusions",
+                    "categoryExclusions"
+                  ],
+                  "merchantExclusions": [
+                    "merchantExclusions",
+                    "merchantExclusions"
+                  ],
+                  "offerFeatures": [
+                    "INTERACTIVE",
+                    "INTERACTIVE"
+                  ]
+                },
                 "categories": [
                   "Arts & Entertainment",
                   "Arts & Entertainment"