+++ src/KardFinancial.Test/Unit/MockServer/Organizations/ContentStrategies/ListTest.cs
@@ -22,6 +22,24 @@ public class ListTest : BaseMockServerTest
                     "name": "name",
                     "organizationId": "organizationId",
                     "sort": "NEWLY_LIVE",
+                    "filters": {
+                      "categories": [
+                        "categories",
+                        "categories"
+                      ],
+                      "categoryExclusions": [
+                        "categoryExclusions",
+                        "categoryExclusions"
+                      ],
+                      "merchantExclusions": [
+                        "merchantExclusions",
+                        "merchantExclusions"
+                      ],
+                      "offerFeatures": [
+                        "INTERACTIVE",
+                        "INTERACTIVE"
+                      ]
+                    },
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"
@@ -43,6 +61,24 @@ public class ListTest : BaseMockServerTest
                     "name": "name",
                     "organizationId": "organizationId",
                     "sort": "NEWLY_LIVE",
+                    "filters": {
+                      "categories": [
+                        "categories",
+                        "categories"
+                      ],
+                      "categoryExclusions": [
+                        "categoryExclusions",
+                        "categoryExclusions"
+                      ],
+                      "merchantExclusions": [
+                        "merchantExclusions",
+                        "merchantExclusions"
+                      ],
+                      "offerFeatures": [
+                        "INTERACTIVE",
+                        "INTERACTIVE"
+                      ]
+                    },
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"