+++ src/KardFinancial.Test/Unit/MockServer/Organizations/Placements/ListTest.cs
@@ -60,6 +60,24 @@ public class ListTest : BaseMockServerTest
                     "name": "name",
                     "organizationId": "organizationId",
                     "sort": "NEWLY_LIVE",
+                    "filters": {
+                      "categories": [
+                        "categories",
+                        "categories"
+                      ],
+                      "categoryExclusions": [
+                        "categoryExclusions",
+                        "categoryExclusions"
+                      ],
+                      "merchantExclusions": [
+                        "merchantExclusions",
+                        "merchantExclusions"
+                      ],
+                      "offerFeatures": [
+                        "INTERACTIVE",
+                        "INTERACTIVE"
+                      ]
+                    },
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"
@@ -81,6 +99,24 @@ public class ListTest : BaseMockServerTest
                     "name": "name",
                     "organizationId": "organizationId",
                     "sort": "NEWLY_LIVE",
+                    "filters": {
+                      "categories": [
+                        "categories",
+                        "categories"
+                      ],
+                      "categoryExclusions": [
+                        "categoryExclusions",
+                        "categoryExclusions"
+                      ],
+                      "merchantExclusions": [
+                        "merchantExclusions",
+                        "merchantExclusions"
+                      ],
+                      "offerFeatures": [
+                        "INTERACTIVE",
+                        "INTERACTIVE"
+                      ]
+                    },
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"