+++ src/KardFinancial.Test/Unit/Serialization/CreateContentStrategyAttributesTest.cs
@@ -17,6 +17,11 @@ public class CreateContentStrategyAttributesTest
             {
               "name": "Featured Travel",
               "sort": "HIGHEST_CASHBACK",
+              "filters": {
+                "offerFeatures": [
+                  "INTERACTIVE"
+                ]
+              },
               "categories": [
                 "Travel"
               ],
@@ -32,6 +37,10 @@ public class CreateContentStrategyAttributesTest
         {
             Name = "Featured Travel",
             Sort = ContentStrategySort.HighestCashback,
+            Filters = new ContentStrategyFilters
+            {
+                OfferFeatures = new List<OfferFeatures>() { OfferFeatures.Interactive },
+            },
             Categories = new List<CategoryOption>() { CategoryOption.Travel },
             CategoryExclusions = new List<CategoryOption>() { CategoryOption.Gas },
             MerchantExclusions = new List<string>() { "merchant-abc" },
@@ -47,6 +56,11 @@ public class CreateContentStrategyAttributesTest
             {
               "name": "Featured Travel",
               "sort": "HIGHEST_CASHBACK",
+              "filters": {
+                "offerFeatures": [
+                  "INTERACTIVE"
+                ]
+              },
               "categories": [
                 "Travel"
               ],