+++ src/KardFinancial/Organizations/ContentStrategies/Types/CreateContentStrategyAttributes.cs
@@ -28,6 +28,12 @@ public record CreateContentStrategyAttributes : IJsonOnDeserialized
     public ContentStrategySort? Sort { get; set; }
 
     /// <summary>
+    /// Filters applied when selecting offers for the strategy
+    /// </summary>
+    [JsonPropertyName("filters")]
+    public required ContentStrategyFilters Filters { get; set; }
+
+    /// <summary>
     /// Merchant categories to include
     /// </summary>
     [JsonPropertyName("categories")]